+++ package.json
@@ -10,9 +10,11 @@
     "jasmine-core": "^2.4.1",
     "jsdoc": "^3.3.0-alpha9",
     "jshint": "~2.8.0",
+    "mkdirp": "^0.5.1",
     "node-ensure": "^0.0.0",
     "requirejs": "^2.1.22",
     "rimraf": "^2.4.1",
+    "run-sequence": "^1.2.2",
     "shelljs": "~0.4.0",
     "streamqueue": "^1.1.1",
     "typogr": "~0.6.5",
